pyproject.toml
[project]
name = "hnh"
version = "0.2.0"
description = "hotdog not hotdog"
authors = [
    {name = "Mingk42", email = "xoals123456t@gmail.com"},
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.6",
    "transformers>=4.45.0",
    "pillow>=10.4.0",
    "python-multipart>=0.0.10",
    "torch>=2.4.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pdm.dev-dependencies]
note = [
    "notebook>=7.2.2",
]
test = [
    "pytest>=8.3.3",
]
